refactor(canvas): deepen chat page rail and tools
